refactor: move all speech related code to feature directory
