+++ lib/features/journal/ui/widgets/entry_details_widget.dart
@@ -13,8 +13,8 @@ import 'package:lotti/features/journal/ui/widgets/entry_details/workout_summary.
 import 'package:lotti/features/journal/ui/widgets/entry_image_widget.dart';
 import 'package:lotti/features/journal/ui/widgets/journal_card.dart';
 import 'package:lotti/features/journal/ui/widgets/tags/tags_list_widget.dart';
+import 'package:lotti/features/speech/ui/widgets/audio_player.dart';
 import 'package:lotti/features/tasks/ui/task_form.dart';
-import 'package:lotti/widgets/audio/audio_player.dart';
 import 'package:lotti/widgets/events/event_form.dart';
 
 class EntryDetailWidget extends ConsumerWidget {
