+++ lib/features/speech/state/player_state.dart
