+++ lib/features/speech/state/player_state.freezed.dart
