+++ lib/features/speech/state/recorder_state.freezed.dart
