+++ lib/features/speech/state/speech_settings_state.freezed.dart
