+++ lib/features/speech/ui/pages/record_audio_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:lotti/features/speech/ui/widgets/audio_recorder.dart';
 import 'package:lotti/l10n/app_localizations_context.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
-import 'package:lotti/widgets/audio/audio_recorder.dart';
 
 class RecordAudioPage extends StatelessWidget {
   const RecordAudioPage({
