+++ lib/features/speech/ui/pages/speech_settings_page.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:lotti/blocs/settings/speech/speech_settings_cubit.dart';
-import 'package:lotti/blocs/settings/speech/speech_settings_state.dart';
 import 'package:lotti/database/maintenance.dart';
+import 'package:lotti/features/speech/state/speech_settings_cubit.dart';
+import 'package:lotti/features/speech/state/speech_settings_state.dart';
+import 'package:lotti/features/speech/ui/widgets/whisper_model_card.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/l10n/app_localizations_context.dart';
 import 'package:lotti/pages/settings/sliver_box_adapter_page.dart';
 import 'package:lotti/widgets/settings/settings_card.dart';
-import 'package:lotti/widgets/settings/speech/whisper_model_card.dart';
 
 class SpeechSettingsPage extends StatelessWidget {
   const SpeechSettingsPage({super.key});
