+++ lib/features/speech/ui/widgets/transcription_progress_modal.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
+import 'package:lotti/features/speech/state/asr_service.dart';
 import 'package:lotti/get_it.dart';
-import 'package:lotti/services/asr_service.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/misc/wolt_modal_config.dart';
 import 'package:wolt_modal_sheet/wolt_modal_sheet.dart';
