+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: Achieve your goals and keep your data private with Lotti.
 publish_to: 'none'
-version: 0.9.512+2683
+version: 0.9.513+2684
 
 msix_config:
   display_name: LottiApp
