+++ test/blocs/settings/speech/speech_settings_cubit_test.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/database/logging_db.dart';
 import 'package:lotti/database/settings_db.dart';
+import 'package:lotti/features/speech/state/asr_service.dart';
 import 'package:lotti/get_it.dart';
-import 'package:lotti/services/asr_service.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
