+++ test/widgets/audio/audio_recording_indicator_test.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:lotti/blocs/audio/recorder_cubit.dart';
-import 'package:lotti/blocs/audio/recorder_state.dart';
 import 'package:lotti/database/settings_db.dart';
+import 'package:lotti/features/speech/state/recorder_cubit.dart';
+import 'package:lotti/features/speech/state/recorder_state.dart';
+import 'package:lotti/features/speech/ui/widgets/audio_recording_indicator.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/nav_service.dart';
-import 'package:lotti/widgets/audio/audio_recording_indicator.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../mocks/mocks.dart';
